+++ lib/data/repositories/open_ai_repository.dart
@@ -1,6 +1,10 @@
+import 'dart:io';
+
 import 'package:dio/dio.dart';
 import 'package:open_ai_simplified/data/remote/open_ia_service.dart';
 import 'package:open_ai_simplified/domain/exceptions.dart';
+import 'package:open_ai_simplified/domain/models/config_images.dart';
+import 'package:open_ai_simplified/domain/models/images_response.dart';
 import 'package:open_ai_simplified/domain/models/models.dart';
 
 class OpenIARepository {
@@ -8,16 +12,70 @@ class OpenIARepository {
     OpenIAService? service,
   }) : _service = service ?? OpenIAService(Dio());
 
-  ConfigCompletion _configCompletion = ConfigCompletion();
+  // private that contains the apiKey of the user need to be initialized via addApiKey method
   String _apiKey = '';
+
+  // contains a configCompletion object that have the information necessary to configure the
+  // service of getting completions, default values are setted but you can change it with
+  // the method configCompletionFromMap or configCompletionFromConfig
+  ConfigCompletion _configCompletion = ConfigCompletion();
+
+  // contains a ConfigEdits object that have the information necessary to configure the
+  // service of getting edits, default values are setted but you can change it with
+  // the method configEditsFromMap or configEditsFromConfig
   ConfigEdits _configEdits = ConfigEdits();
+
+  // contains a ConfigImages object that have the information necessary to configure the
+  // service of getting images, default values are setted but you can change it with
+  // the method configImagesFromMap or configImagesFromConfig
+  ConfigImages _configImages = ConfigImages();
+
+  // this is the service that makes tha API calls
   OpenIAService get service => _service;
   final OpenIAService _service;
 
+  // Add an APIKey to the package
   void addApiKey(String apiKey) {
     _apiKey = apiKey;
   }
 
+  // configure the images service from a map
+  void configImagesFromMap(Map<String, dynamic> newConfig) {
+    if (newConfig['n'] != null && (newConfig['n'] < 1 || newConfig['n'] > 10)) {
+      throw InvalidParamsException(message: 'the n param is lower than 1');
+    }
+    if (newConfig['size'] != '256x256' ||
+        newConfig['size'] != '512x512' ||
+        newConfig['size'] != '1024x1024') {
+      throw InvalidParamsException(
+          message:
+              'the size param must be ne of 256x256, 512x512, or 1024x1024');
+    }
+    _configImages = _configImages.copyWith(
+      n: newConfig['n'] ?? _configImages.n,
+      size: newConfig['size'] ?? _configImages.size,
+    );
+  }
+
+  // configure the images service from a ConfigImages object
+  void configImagesFromConfig(ConfigImages newConfig) {
+    if (newConfig.n < 1 || newConfig.n > 10) {
+      throw InvalidParamsException(message: 'the n param is lower than 1');
+    }
+    if (newConfig.size != '256x256' &&
+        newConfig.size != '512x512' &&
+        newConfig.size != '1024x1024') {
+      throw InvalidParamsException(
+          message:
+              'the size param must be ne of 256x256, 512x512, or 1024x1024');
+    }
+    _configImages = _configImages.copyWith(
+      n: newConfig.n,
+      size: newConfig.size,
+    );
+  }
+
+  // configure the completion service from a map
   void configCompletionFromMap(Map<String, dynamic> newConfig) {
     _configCompletion = _configCompletion.copyWith(
         maxTokens: newConfig['max_token'] ?? _configCompletion.maxTokens,
@@ -30,6 +88,7 @@ class OpenIARepository {
         stop: newConfig['stop'] ?? _configCompletion);
   }
 
+  // configure the completion service from a ConfigCompletion object
   void configCompletionFromConfig(ConfigCompletion config) {
     _configCompletion = _configCompletion.copyWith(
       maxTokens: config.maxTokens,
@@ -43,6 +102,7 @@ class OpenIARepository {
     );
   }
 
+  // configure the edits service from a map
   void configEditsFromMap({
     required Map<String, dynamic> newConfig,
     required bool isText,
@@ -66,6 +126,7 @@ class OpenIARepository {
     );
   }
 
+  // configure the edits service from a ConfigEdits object
   void configEditsFromConfig({
     required ConfigEdits config,
   }) {
@@ -88,32 +149,49 @@ class OpenIARepository {
     );
   }
 
+  // get a Completion as Map<String, dynamic>
   Future<Map<String, dynamic>> getRawCompletion(String prompt) async {
     try {
       if (_apiKey.isEmpty) {
         throw KeyNotFoundException();
       }
+      if (prompt.isEmpty) {
+        throw InvalidParamsException();
+      }
       final response = await service.getCompletion(
           prompt: prompt, apiKey: _apiKey, config: _configCompletion);
       return response.toMap();
     } catch (e) {
-      throw Exception(e.toString());
+      if (e is OpenAIException) {
+        rethrow;
+      } else {
+        throw Exception(e.toString());
+      }
     }
   }
 
+  // get a completion as CompletionResponse object
   Future<CompletionResponse> getCompletion(String prompt) async {
     try {
       if (_apiKey.isEmpty) {
         throw KeyNotFoundException();
       }
+      if (prompt.isEmpty) {
+        throw InvalidParamsException();
+      }
       final response = await service.getCompletion(
           prompt: prompt, apiKey: _apiKey, config: _configCompletion);
       return response;
     } catch (e) {
-      throw Exception(e.toString());
+      if (e is OpenAIException) {
+        rethrow;
+      } else {
+        throw Exception(e.toString());
+      }
     }
   }
 
+  // get a Models as Map<String, dynamic>
   Future<Map<String, dynamic>> getRawModelsList() async {
     try {
       if (_apiKey.isEmpty) {
@@ -122,10 +200,15 @@ class OpenIARepository {
       final response = await service.getModelsList(apiKey: _apiKey);
       return response.toMap();
     } catch (e) {
-      throw Exception(e.toString());
+      if (e is OpenAIException) {
+        rethrow;
+      } else {
+        throw Exception(e.toString());
+      }
     }
   }
 
+  // get the Models as OpenAiModels object
   Future<OpenAiModels> getModelsList() async {
     try {
       if (_apiKey.isEmpty) {
@@ -134,10 +217,15 @@ class OpenIARepository {
       final response = await service.getModelsList(apiKey: _apiKey);
       return response;
     } catch (e) {
-      throw Exception(e.toString());
+      if (e is OpenAIException) {
+        rethrow;
+      } else {
+        throw Exception(e.toString());
+      }
     }
   }
 
+  // get a Edit as Map<String, dynamic>
   Future<Map<String, dynamic>> getRawEdits(
       {required String input, required String instruction}) async {
     try {
@@ -152,10 +240,15 @@ class OpenIARepository {
           apiKey: _apiKey, config: _configEdits, inputWithInstruction: map);
       return response.toMap();
     } catch (e) {
-      rethrow;
+      if (e is OpenAIException) {
+        rethrow;
+      } else {
+        throw Exception(e.toString());
+      }
     }
   }
 
+  // get a Edit as EditsResponse object
   Future<EditsResponse> getEdits(
       {required String input, required String instruction}) async {
     try {
@@ -170,7 +263,154 @@ class OpenIARepository {
           apiKey: _apiKey, config: _configEdits, inputWithInstruction: map);
       return response;
     } catch (e) {
-      rethrow;
+      if (e is OpenAIException) {
+        rethrow;
+      } else {
+        throw Exception(e.toString());
+      }
+    }
+  }
+
+  // get Images as Map<String, dynamic>
+  Future<Map<String, dynamic>> getRawImages(String prompt) async {
+    try {
+      if (_apiKey.isEmpty) {
+        throw KeyNotFoundException();
+      }
+      if (prompt.isEmpty) {
+        throw InvalidParamsException();
+      }
+      final map = {'prompt': prompt};
+      final result = await service.generateImages(
+          apiKey: _apiKey, config: _configImages, prompt: map);
+      return result.toMap();
+    } catch (e) {
+      if (e is OpenAIException) {
+        rethrow;
+      } else {
+        throw Exception(e.toString());
+      }
+    }
+  }
+
+  // get Images as ImagesResponse object
+  Future<ImagesResponse> getImages(String prompt) async {
+    try {
+      if (_apiKey.isEmpty) {
+        throw KeyNotFoundException();
+      }
+      if (prompt.isEmpty) {
+        throw InvalidParamsException();
+      }
+      final map = {'prompt': prompt};
+      final result = await service.generateImages(
+          apiKey: _apiKey, config: _configImages, prompt: map);
+      return result;
+    } catch (e) {
+      if (e is OpenAIException) {
+        rethrow;
+      } else {
+        throw Exception(e.toString());
+      }
+    }
+  }
+
+  // create a variation of an image and return as ImagesResponse object
+  Future<ImagesResponse> createAImageVariation(
+      {required File imageFile}) async {
+    try {
+      if (_apiKey.isEmpty) {
+        throw KeyNotFoundException();
+      }
+
+      final result = await service.variateImage(
+          image: imageFile, apiKey: _apiKey, config: _configImages);
+      return result;
+    } catch (e) {
+      if (e is OpenAIException) {
+        rethrow;
+      } else {
+        throw Exception(e.toString());
+      }
+    }
+  }
+
+  // create a variation of an image and return as a Map
+  Future<Map<String, dynamic>> createRawImageVariation(
+      {required File imageFile}) async {
+    try {
+      if (_apiKey.isEmpty) {
+        throw KeyNotFoundException();
+      }
+
+      final result = await service.variateImage(
+          image: imageFile, apiKey: _apiKey, config: _configImages);
+      return result.toMap();
+    } catch (e) {
+      if (e is OpenAIException) {
+        rethrow;
+      } else {
+        throw Exception(e.toString());
+      }
+    }
+  }
+
+  // Creates an edited or extended image given an original image and a prompt.
+  // Provide a mask is optional
+  // An mask is: An additional image whose fully transparent areas (e.g. where alpha is zero) indicate where image should be edited. Must be a valid PNG file, less than 4MB, and have the same dimensions as image.
+  // The image to edit. Must be a valid PNG file, less than 4MB, and square. If mask is not provided, image must have transparency, which will be used as the mask.
+  Future<ImagesResponse> editImage({
+    required File image,
+    File? mask,
+    required String prompt,
+  }) async {
+    try {
+      if (_apiKey.isEmpty) {
+        throw KeyNotFoundException();
+      }
+      if (prompt.isEmpty) {
+        throw InvalidParamsException();
+      }
+      final result = await service.editImage(
+          prompt: prompt,
+          image: image,
+          apiKey: _apiKey,
+          config: _configImages,
+          mask: mask);
+      return result;
+    } catch (e) {
+      if (e is OpenAIException) {
+        rethrow;
+      } else {
+        throw Exception(e.toString());
+      }
+    }
+  }
+
+  // Creates an edited or extended image given an original image and a prompt.
+  // Provide a mask is optional
+  // An mask is: An additional image whose fully transparent areas (e.g. where alpha is zero) indicate where image should be edited. Must be a valid PNG file, less than 4MB, and have the same dimensions as image.
+  Future<Map<String, dynamic>> editRawImage({
+    required File image,
+    File? mask,
+    required String prompt,
+  }) async {
+    try {
+      if (_apiKey.isEmpty) {
+        throw KeyNotFoundException();
+      }
+      if (prompt.isEmpty) {
+        throw InvalidParamsException();
+      }
+      final result = await service.editImage(
+          prompt: prompt, image: image, apiKey: _apiKey, config: _configImages);
+      return result.toMap();
+    } catch (e) {
+      if (e is OpenAIException) {
+        rethrow;
+      } else {
+        throw Exception(e.toString());
+      }
     }
   }
 }