+++ lib/domain/models/completion_response.dart
@@ -1,14 +1,17 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'dart:convert';
-
 import 'package:open_ai_simplified/domain/models/usage.dart';
 
 class CompletionResponse {
+  // id of the object
   final String id;
+  // type of the object
   final String object;
+  // date of the moment were the object was created
   final int created;
+  // model utilized to create the completion
   final String model;
+  // list of the values that brings the completion
   final List<Choices> choices;
+  // object that shows how much Tokens were spended on the request + response
   final Usage usage;
 
   CompletionResponse(
@@ -19,6 +22,7 @@ class CompletionResponse {
       required this.choices,
       required this.usage});
 
+  // Converts the object to a Map<String, dynamic>
   Map<String, dynamic> toMap() {
     return <String, dynamic>{
       'id': id,
@@ -30,6 +34,7 @@ class CompletionResponse {
     };
   }
 
+  // Create the object from a Map<String, dynamic>
   factory CompletionResponse.fromMap(Map<String, dynamic> map) {
     return CompletionResponse(
       id: map['id'] as String,
@@ -44,17 +49,16 @@ class CompletionResponse {
       usage: Usage.fromMap(map['usage'] as Map<String, dynamic>),
     );
   }
-
-  String toJson() => json.encode(toMap());
-
-  factory CompletionResponse.fromJson(String source) =>
-      CompletionResponse.fromMap(json.decode(source) as Map<String, dynamic>);
 }
 
 class Choices {
+  // Text requested
   final String text;
+  // Index of the present text on the list
   final int index;
+  // Include the log probabilities on the logprobs most likely tokens, as well the chosen tokens. For example, if logprobs is 5, the API will return a list of the 5 most likely tokens. The API will always return the logprob of the sampled token, so there may be up to logprobs+1 elements in the response.
   final int? logprobs;
+  // In case of the response need to stop, here will be the awnser to why it stop.
   final String finishReason;
 
   Choices({
@@ -64,6 +68,7 @@ class Choices {
     required this.finishReason,
   });
 
+  // Converts the object to a Map<String, dynamic>
   Map<String, dynamic> toMap() {
     return <String, dynamic>{
       'text': text,
@@ -73,6 +78,7 @@ class Choices {
     };
   }
 
+  // Create the object from a Map<String, dynamic>
   factory Choices.fromMap(Map<String, dynamic> map) {
     return Choices(
       text: map['text'] as String,
@@ -81,9 +87,4 @@ class Choices {
       finishReason: map['finish_reason'] as String,
     );
   }
-
-  String toJson() => json.encode(toMap());
-
-  factory Choices.fromJson(String source) =>
-      Choices.fromMap(json.decode(source) as Map<String, dynamic>);
 }