+++ lib/domain/models/edits_response.dart
@@ -1,16 +1,18 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'dart:convert';
-
 import 'package:open_ai_simplified/domain/models/usage.dart';
 
 class EditsResponse {
+  // type of the object
   final String object;
+  // date of the moment were the object was created
   final int created;
+  // list of the values that brings the Edit
   final List<ChoicesEdit> choices;
+  // object that shows how much Tokens were spended on the request + response
   final Usage usage;
 
   EditsResponse(this.object, this.created, this.choices, this.usage);
 
+  // Converts the object to a Map<String, dynamic>
   Map<String, dynamic> toMap() {
     return <String, dynamic>{
       'object': object,
@@ -20,6 +22,7 @@ class EditsResponse {
     };
   }
 
+  // Create the object from a Map<String, dynamic>
   factory EditsResponse.fromMap(Map<String, dynamic> map) {
     return EditsResponse(
       map['object'] as String,
@@ -32,19 +35,17 @@ class EditsResponse {
       Usage.fromMap(map['usage'] as Map<String, dynamic>),
     );
   }
-
-  String toJson() => json.encode(toMap());
-
-  factory EditsResponse.fromJson(String source) =>
-      EditsResponse.fromMap(json.decode(source) as Map<String, dynamic>);
 }
 
 class ChoicesEdit {
+  // the text edited
   final String text;
+  // index on de list of texts editeds
   final int index;
 
   ChoicesEdit(this.text, this.index);
 
+  // Converts the object to a Map<String, dynamic>
   Map<String, dynamic> toMap() {
     return <String, dynamic>{
       'text': text,
@@ -52,15 +53,11 @@ class ChoicesEdit {
     };
   }
 
+  // Create the object from a Map<String, dynamic>
   factory ChoicesEdit.fromMap(Map<String, dynamic> map) {
     return ChoicesEdit(
       map['text'] as String,
       map['index'] as int,
     );
   }
-
-  String toJson() => json.encode(toMap());
-
-  factory ChoicesEdit.fromJson(String source) =>
-      ChoicesEdit.fromMap(json.decode(source) as Map<String, dynamic>);
 }