+++ lib/domain/models/open_ai_models.dart
@@ -1,17 +1,17 @@
-import 'dart:convert';
-
-// ignore_for_file: public_member_api_docs, sort_constructors_first
 class OpenAiModels {
+  // List with the information about models
   final List<Data> data;
 
   OpenAiModels(this.data);
 
+  // Converts the object to a Map<String, dynamic>
   Map<String, dynamic> toMap() {
     return <String, dynamic>{
       'data': data.map((x) => x.toMap()).toList(),
     };
   }
 
+  // Create the object from a Map<String, dynamic>
   factory OpenAiModels.fromMap(Map<String, dynamic> map) {
     return OpenAiModels(
       List<Data>.from(
@@ -21,16 +21,14 @@ class OpenAiModels {
       ),
     );
   }
-
-  String toJson() => json.encode(toMap());
-
-  factory OpenAiModels.fromJson(String source) =>
-      OpenAiModels.fromMap(json.decode(source) as Map<String, dynamic>);
 }
 
 class Data {
+  // id of the model [used on the configs]
   final String? id;
+  // type of object
   final String? objec;
+  // owner of the model
   final String? ownedBy;
   Data({
     this.id,
@@ -38,6 +36,7 @@ class Data {
     this.ownedBy,
   });
 
+  // Converts the object to a Map<String, dynamic>
   Map<String, dynamic> toMap() {
     return <String, dynamic>{
       'id': id,
@@ -46,6 +45,7 @@ class Data {
     };
   }
 
+  // Create the object from a Map<String, dynamic>
   factory Data.fromMap(Map<String, dynamic> map) {
     return Data(
       id: map['id'] != null ? map['id'] as String : null,
@@ -53,9 +53,4 @@ class Data {
       ownedBy: map['ownedBy'] != null ? map['ownedBy'] as String : null,
     );
   }
-
-  String toJson() => json.encode(toMap());
-
-  factory Data.fromJson(String source) =>
-      Data.fromMap(json.decode(source) as Map<String, dynamic>);
 }